requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
chromadb>=0.5.0
tiktoken>=0.7.0
pydantic>=2.0.0
pypdf>=5.0.0
google-genai>=1.0.0
Pillow>=10.0.0
